+++ packages/elfutils.rb
@@ -27,7 +27,7 @@ class Elfutils < Package
   depends_on 'gcc' # R
   depends_on 'glibc' # R
   depends_on 'libarchive' # R
-  depends_on 'libcurl' # R
+  depends_on 'curl' # R
   depends_on 'libmicrohttpd' # R
   depends_on 'sqlite' # R
   depends_on 'xzutils' # R
