+++ packages/grive.rb
@@ -28,7 +28,7 @@ class Grive < Package
   depends_on 'expat' => :build
   depends_on 'gcc' # R
   depends_on 'glibc' # R
-  depends_on 'libcurl' # R
+  depends_on 'curl' # R
   depends_on 'libgcrypt' # R
   depends_on 'libgpgerror' # R
   depends_on 'yajl' # R
