+++ packages/libexiv2.rb
@@ -28,7 +28,7 @@ class Libexiv2 < Package
   depends_on 'expat' # R
   depends_on 'gcc' # R
   depends_on 'glibc' # R
-  depends_on 'libcurl' # R
+  depends_on 'curl' # R
   depends_on 'zlibpkg' # R
 
   def self.patch
