+++ packages/mupdf.rb
@@ -27,7 +27,7 @@ class Mupdf < Package
   depends_on 'gcc' # R
   depends_on 'glibc' # R
   depends_on 'jbigkit'
-  depends_on 'libcurl' # R
+  depends_on 'curl' # R
   depends_on 'libjpeg'
   depends_on 'libx11' # R
   depends_on 'libxext' # R
