+++ packages/php5.rb
@@ -26,7 +26,7 @@ class Php5 < Package
   depends_on 'libpng'
   depends_on 'libxslt'
   depends_on 'libzip'
-  depends_on 'libcurl'
+  depends_on 'curl'
   depends_on 'exif'
   depends_on 'freetype'
   depends_on 'pcre'
