+++ packages/php71.rb
@@ -27,7 +27,7 @@ class Php71 < Package
   depends_on 'libwebp'
   depends_on 'libxslt'
   depends_on 'libzip'
-  depends_on 'libcurl'
+  depends_on 'curl'
   depends_on 'exif'
   depends_on 'freetype'
   depends_on 'pcre'
