+++ packages/php74.rb
@@ -24,7 +24,7 @@ class Php74 < Package
   })
 
   depends_on 'aspell_en'
-  depends_on 'libcurl'
+  depends_on 'curl'
   depends_on 'libgcrypt'
   depends_on 'libjpeg'
   depends_on 'libpng'
