+++ packages/qemu.rb
@@ -54,7 +54,7 @@ class Qemu < Package
   depends_on 'gcc' # R
   depends_on 'glibc' # R
   depends_on 'gnutls' # R
-  depends_on 'libcurl' # R
+  depends_on 'curl' # R
   depends_on 'libcyrussasl' # R
   depends_on 'libpng' # R
   depends_on 'libseccomp' # R
