+++ packages/weechat.rb
@@ -24,7 +24,7 @@ class Weechat < Package
 
   depends_on 'libgcrypt'
   depends_on 'aspell' => :build
-  depends_on 'libcurl' => :build
+  depends_on 'curl' => :build
   depends_on 'lua' => :build
   depends_on 'python2'
   depends_on 'tcl' => :build
